refactor: move store disk-read and merge helpers into store-codec.ts
Relocate readDiskTasks and mergeWithDisk from TaskIndexStore to the existing
stateless store-codec.ts module. The functions are moved verbatim with only
signature changes (path arguments instead of this.* access). store.ts drops
from 498 to 432 lines, keeping both files well under the 500-line cap.

No behavioral change: doSave's lock critical section, backup-on-corrupt, and
the three-way merge invariants (dirty wins, removed stays removed, peer
creates are preserved) are preserved exactly.
